[fix]ログインによる機能制限

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
@@ -225,7 +226,10 @@ class UserController extends Controller
     public function deleteData(Request $request)
     {
         $user = User::find($request->id);
-        $user->delete();
+
+        if (Auth::id() === $user->id) {
+            $user->delete();
+        }
 
         return $this->resigned($request, $user)
             ?: redirect($this->redirectPath());
